Separate llm-generation logic
